Added pyIslam
I added another python repository. I also changed the name of the repository pyIslam to Islam. I did this because the pyIslam(abougouffa/pyIslam) package name is actually islam in pypi, while for the other pyIslam (Moha369/PyIslam) it is the actual pyIslam. This is also mentioned in the pyIslam(abougouffa/pyIslam) repo as well.