Added info on whether project is continued or discontinued
For each project, I added a tick or cross, to symbolize whether the project is opened or closed. So if for example the last commit was 9 years ago, it is probably not being worked on. But if the last commit last less than a year ago, it might still have been being worked on.